refactor: drop passthrough barrels in integrations
Same shape as @promster/metrics: src/<name>/index.ts re-exporting
src/<name>/<name>.ts, one implementation per directory. Flatten to
src/<name>.ts and delete the six barrels. Entry barrels and spec
imports need no change, since './plugin' resolves to the file it
previously resolved to as a directory.

hapi and fastify additionally needed their package.json import fixed.
Both read the plugin name and version from '../../package.json', which
escaped src/ by exactly the level the flattening removed.
