--- packages/apollo/src/plugin/index.ts
@@ -1,8 +1,0 @@
-import {
-  createPlugin,
-  signalIsNotUp,
-  signalIsUp,
-  type TPromsterOptions,
-} from './plugin';
-
-export { type TPromsterOptions, createPlugin, signalIsUp, signalIsNotUp };
