--- packages/express/src/middleware/index.ts
@@ -1,15 +1,0 @@
-import {
-  createMiddleware,
-  getRequestRecorder,
-  signalIsNotUp,
-  signalIsUp,
-  type TPromsterOptions,
-} from './middleware';
-
-export {
-  type TPromsterOptions,
-  createMiddleware,
-  signalIsUp,
-  signalIsNotUp,
-  getRequestRecorder,
-};
