--- packages/fastify/src/plugin/index.ts
@@ -1,15 +1,0 @@
-import {
-  getRequestRecorder,
-  plugin,
-  signalIsNotUp,
-  signalIsUp,
-  type TPromsterOptions,
-} from './plugin';
-
-export {
-  type TPromsterOptions,
-  plugin,
-  signalIsUp,
-  signalIsNotUp,
-  getRequestRecorder,
-};
