+++ packages/hapi/src/plugin.spec.js
