--- packages/hapi/src/plugin/index.ts
@@ -1,15 +1,0 @@
-import {
-  createPlugin,
-  getRequestRecorder,
-  signalIsNotUp,
-  signalIsUp,
-  type TPromsterOptions,
-} from './plugin';
-
-export {
-  type TPromsterOptions,
-  createPlugin,
-  getRequestRecorder,
-  signalIsUp,
-  signalIsNotUp,
-};
