--- packages/server/src/server/index.ts
@@ -1,3 +1,0 @@
-import { createServer } from './server';
-
-export { createServer };
